PR S4.4.5: extract views/diagnostic/ from diagnostic_cog
Applies the F-3 view-extraction pattern to the diagnostic subsystem.
Cog drops 913 → 763 LOC, now under S4.6's 800-LOC fail threshold.
Same shape as the prior S4.2-followup xp and S4.3 moderation passes;
no PersistentView here (both diagnostic views are ephemeral, timeout-
based), so Pattern A applies — no Pattern-B re-export needed.

Why this stage, not S4.5 (blackjack):

The original audit row A4 (blackjack_cog decomposition) was assessed
and found to require deeper test-invariant migration than fits a
controlled-mode mini-release.  blackjack_cog has 12+ tests that
import module-level persistence functions (_save_solo_game,
_clear_solo_game, _save_pvp_match, _clear_pvp_match, etc.) and patch
cogs.blackjack_cog.game_state_service.save directly.  Moving the
persistence helpers to a sub-package would force a test-update PR
bundled with the move.  That work is real but it's its own stage.

S4.4.5 (this stage) is the audit's addendum §10.3 follow-on: the
diagnostic_cog was identified as the cog that would otherwise FAIL
S4.6's size invariant after the rest of Phase A landed.  Doing it
now unblocks S4.6 and follows the canonical view-extraction pattern
established by S4.2-followup and S4.3.

New layout (mirrors views/economy/, views/moderation/):

  views/diagnostic/__init__.py    — re-exports
  views/diagnostic/paginator.py   — _PaginatorView (BaseView)
  views/diagnostic/hub_panel.py   — _DiagnosticsHubView (HubView)

Behavior preserved end-to-end:

  * !diagnostics opens the same hub panel with the same 8 buttons,
    each one ctx.invoke-ing the corresponding command.
  * !platform <subsystem> group commands all work — they live on
    DiagnosticCog, not in the extracted views.
  * Help-menu direct navigation (build_help_menu_view) still returns
    a working _DiagnosticsHubView via help_ctx_shim.
  * _PaginatorView is used by list_commands_detailed; same call site,
    same behavior.
  * No PersistentView in this subsystem, so identity contract surface
    unchanged (no @register, no panel_anchors row).

Cog-side cleanup:

  * Removed unused imports: BaseView, HubView (moved with the views),
    safe_defer (only the extracted hub view used it).
  * send_panel and help_ctx_shim retained — still consumed by the
    diagnostics_hub command and build_help_menu_view hook.

Validation:

  * pytest tests/unit/ → 837 passed, 0 failed
  * test_platform_commands.py + test_platform_diagnostics_commands.py
    pass unchanged (they reference DiagnosticCog, not the views).
  * Identity contract STRICT mode green
  * INV-A..INV-N green
  * black, isort, ruff, mypy all clean over the whole disbot/ tree
    (168 source files)

Cogs over 800 LOC after this stage:

  - blackjack_cog.py     1702 LOC  ⚠ S4.5 work
  - rps_tournament_cog.py 1182 LOC  ⚠ further extraction needed
  - role_cog.py            620 LOC  warn-tier (acceptable per audit §3.5)
  - counting_cog.py        654 LOC  warn-tier (acceptable)
  - moderation_cog.py      212 LOC  ✓ clean (S4.3)
  - help_cog.py            506 LOC  warn-tier (acceptable)
  - chain_cog.py           504 LOC  warn-tier (acceptable)
  - xp_cog.py              110 LOC  ✓ clean (S4.2-followup)
  - diagnostic_cog.py      763 LOC  ✓ now under fail threshold

S4.6 (soft cog size invariant test) can land once blackjack and the
remaining rps_tournament refactor are done.
