+++ disbot/cogs/diagnostic_cog.py
@@ -10,9 +10,10 @@ import discord
 import psutil
 from discord.ext import commands
 
-from core.runtime.interaction_helpers import help_ctx_shim, safe_defer
+from core.runtime.interaction_helpers import help_ctx_shim
 from utils import db
-from views.base import BaseView, HubView, send_panel
+from views.base import send_panel
+from views.diagnostic import _DiagnosticsHubView, _PaginatorView
 
 logger = logging.getLogger("bot")
 
@@ -21,157 +22,6 @@ DATA_DIR = os.path.join(
     "data",
 )
 JSON_DIR = os.path.join(DATA_DIR, "json")
-
-
-class _PaginatorView(BaseView):
-    """Simple prev/next paginator for multi-page embeds."""
-
-    def __init__(
-        self,
-        pages: list[discord.Embed],
-        author: discord.Member | discord.User,
-    ):
-        super().__init__(author, timeout=120)
-        self.pages = pages
-        self.index = 0
-        self._update_buttons()
-
-    def _update_buttons(self):
-        self.prev_btn.disabled = self.index == 0
-        self.next_btn.disabled = self.index == len(self.pages) - 1
-
-    @discord.ui.button(label="◀ Prev", style=discord.ButtonStyle.secondary)
-    async def prev_btn(
-        self,
-        interaction: discord.Interaction,
-        button: discord.ui.Button,
-    ):
-        self.index -= 1
-        self._update_buttons()
-        await interaction.response.edit_message(embed=self.pages[self.index], view=self)
-
-    @discord.ui.button(label="Next ▶", style=discord.ButtonStyle.secondary)
-    async def next_btn(
-        self,
-        interaction: discord.Interaction,
-        button: discord.ui.Button,
-    ):
-        self.index += 1
-        self._update_buttons()
-        await interaction.response.edit_message(embed=self.pages[self.index], view=self)
-
-
-class _DiagnosticsHubView(HubView):
-    """Interactive hub for all diagnostic tools."""
-
-    def __init__(self, ctx: commands.Context, cog: DiagnosticCog):
-        super().__init__(ctx.author)
-        self.ctx = ctx
-        self.cog = cog
-
-    def build_embed(self) -> discord.Embed:
-        embed = discord.Embed(
-            title="🔧 Diagnostics Hub",
-            description=(
-                "Select a diagnostic tool below.\n"
-                "All tools require Administrator permission."
-            ),
-            color=discord.Color.blue(),
-        )
-        embed.add_field(
-            name="🤖 Bot Status",
-            value="Health & performance metrics",
-            inline=True,
-        )
-        embed.add_field(name="📡 Latency", value="WebSocket ping", inline=True)
-        embed.add_field(
-            name="💻 System Info",
-            value="OS, disk & Python version",
-            inline=True,
-        )
-        embed.add_field(
-            name="🗄️ Check Database",
-            value="Verify all DB tables exist",
-            inline=True,
-        )
-        embed.add_field(
-            name="📄 Validate JSON",
-            value="Check data file integrity",
-            inline=True,
-        )
-        embed.add_field(
-            name="📋 Command List",
-            value="Paginated command overview",
-            inline=True,
-        )
-        embed.add_field(
-            name="🔍 Recent Errors",
-            value="Last 10 error log entries",
-            inline=True,
-        )
-        embed.add_field(
-            name="🔔 Test Notify",
-            value="Fire a test webhook ping",
-            inline=True,
-        )
-        embed.set_footer(text="Diagnostics Hub  •  Admin only")
-        return embed
-
-    @discord.ui.button(label="🤖 Bot Status", style=discord.ButtonStyle.blurple, row=0)
-    async def btn_status(self, interaction: discord.Interaction, _: discord.ui.Button):
-        if not await safe_defer(interaction):
-            return
-        await self.ctx.invoke(self.cog.diagnostic_bot_status)
-
-    @discord.ui.button(label="📡 Latency", style=discord.ButtonStyle.blurple, row=0)
-    async def btn_latency(self, interaction: discord.Interaction, _: discord.ui.Button):
-        if not await safe_defer(interaction):
-            return
-        await self.ctx.invoke(self.cog.latency)
-
-    @discord.ui.button(label="💻 System Info", style=discord.ButtonStyle.blurple, row=0)
-    async def btn_sysinfo(self, interaction: discord.Interaction, _: discord.ui.Button):
-        if not await safe_defer(interaction):
-            return
-        await self.ctx.invoke(self.cog.system_info)
-
-    @discord.ui.button(label="🗄️ Database", style=discord.ButtonStyle.grey, row=1)
-    async def btn_db(self, interaction: discord.Interaction, _: discord.ui.Button):
-        if not await safe_defer(interaction):
-            return
-        await self.ctx.invoke(self.cog.check_database)
-
-    @discord.ui.button(label="📄 JSON Files", style=discord.ButtonStyle.grey, row=1)
-    async def btn_json(self, interaction: discord.Interaction, _: discord.ui.Button):
-        if not await safe_defer(interaction):
-            return
-        await self.ctx.invoke(self.cog.validate_json_files)
-
-    @discord.ui.button(label="📋 Commands", style=discord.ButtonStyle.grey, row=1)
-    async def btn_cmds(self, interaction: discord.Interaction, _: discord.ui.Button):
-        if not await safe_defer(interaction):
-            return
-        await self.ctx.invoke(self.cog.list_commands_detailed)
-
-    @discord.ui.button(
-        label="🔍 Recent Errors",
-        style=discord.ButtonStyle.danger,
-        row=2,
-    )
-    async def btn_errors(self, interaction: discord.Interaction, _: discord.ui.Button):
-        if not await safe_defer(interaction):
-            return
-        await self.ctx.invoke(self.cog.recent_errors)
-
-    @discord.ui.button(
-        label="🔔 Test Notify",
-        style=discord.ButtonStyle.secondary,
-        row=2,
-    )
-    async def btn_notify(self, interaction: discord.Interaction, _: discord.ui.Button):
-        if not await safe_defer(interaction):
-            return
-        await self.ctx.invoke(self.cog.test_notification)
 
 
 class DiagnosticCog(commands.Cog):
